refactor general components directory structure. components are scoped by dir/import path rather than by name
