+++ src/components/Interface/ImagesGrid/index.js
@@ -1,0 +1,1 @@
+export { default } from './ImagesGrid'
