+++ src/components/Interface/MobileLinks/index.js
@@ -1,0 +1,1 @@
+export { default } from './MobileLinks'
