+++ src/routes/Interface/Content/Posts/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import InterfacePost from '../../../../components/InterfacePost'
+import Post from '../../../../components/Interface/Post'
 import ContentContainer from '../ContentContainer'
 import withPosts from '../../../../hoc/withPosts'
 
@@ -10,7 +10,7 @@ const Posts = ({ posts }) => (
   <ContentContainer className="interface-posts-container">
     {posts.data.length > 0 &&
       posts.data.map((post, index) =>
-        <InterfacePost
+        <Post
           key={post._id}
           post={post}
           className={`interface-post-${index}`}
