+++ src/routes/Interface/index.js
@@ -14,7 +14,7 @@ import MenuIcon from '@material-ui/icons/Menu'
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft'
 
 import Content from './Content'
-import InterfaceMobileLinks from '../../components/InterfaceMobileLinks'
+import MobileLinks from '../../components/Interface/MobileLinks'
 
 import styles from './styles'
 
@@ -100,7 +100,7 @@ class Interface extends Component {
           <main className={classes.mobileContent}>
             <div className={classes.drawerHeaderPlaceholder} />
             <Content match={match} />
-            <InterfaceMobileLinks />
+            <MobileLinks />
           </main>
         </Hidden>
       </div>
